[refactor] unsed file 삭제, 폴더 이동
