+++ lyc_flutter_project/lib/Join/screens/join_membership_screen_2.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:lyc_flutter_project/data/app_color.dart';
-import 'package:lyc_flutter_project/widget/normal_appbar.dart';
+import 'package:lyc_flutter_project/common/widget/normal_appbar.dart';
 import 'package:lyc_flutter_project/Join/Screens/join_membership_screen_1.dart';
 import 'package:lyc_flutter_project/Join/Screens/join_membership_screen_3.dart';
 import '../../styles/app_text_style.dart';
