+++ lyc_flutter_project/lib/mypage/screen/follow_list_screen.dart
@@ -7,8 +7,8 @@ import 'package:lyc_flutter_project/data/app_color.dart';
 import 'package:lyc_flutter_project/mypage/model/follow_model.dart';
 import 'package:lyc_flutter_project/mypage/provider/follow_provider.dart';
 import 'package:lyc_flutter_project/mypage/repository/mypage_repository.dart';
-import 'package:lyc_flutter_project/styles/default_padding.dart';
-import 'package:lyc_flutter_project/widget/normal_appbar.dart';
+import 'package:lyc_flutter_project/common/widget/default_padding.dart';
+import 'package:lyc_flutter_project/common/widget/normal_appbar.dart';
 import 'package:provider/provider.dart';
 
 class FollowListScreen extends StatelessWidget {
