+++ lyc_flutter_project/lib/posting/screen/add_clothes_posting_screen.dart
@@ -3,11 +3,11 @@ import 'package:image_picker/image_picker.dart';
 import 'package:lyc_flutter_project/common/widget/two_buttons.dart';
 import 'package:lyc_flutter_project/data/app_color.dart';
 import 'package:lyc_flutter_project/posting/provider/clothes_provider.dart';
-import 'package:lyc_flutter_project/styles/posting_text_style.dart';
-import 'package:lyc_flutter_project/widget/image_picker_widget.dart';
-import 'package:lyc_flutter_project/widget/normal_appbar.dart';
-import 'package:lyc_flutter_project/widget/posting_content_text_field.dart';
-import 'package:lyc_flutter_project/widget/select_buttons_in_posting.dart';
+import 'package:lyc_flutter_project/posting/style/posting_text_style.dart';
+import 'package:lyc_flutter_project/common/widget/image_picker_widget.dart';
+import 'package:lyc_flutter_project/common/widget/normal_appbar.dart';
+import 'package:lyc_flutter_project/posting/widget/posting_content_text_field.dart';
+import 'package:lyc_flutter_project/mypage/widget/select_buttons_in_posting.dart';
 import 'package:lyc_flutter_project/common/widget/switch_category_button.dart';
 import 'package:provider/provider.dart';
