+++ lyc_flutter_project/lib/posting/screen/posting_detail_screen.dart
@@ -6,8 +6,8 @@ import 'package:lyc_flutter_project/common/widget/posting_top.dart';
 import 'package:lyc_flutter_project/data/app_color.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:lyc_flutter_project/posting/provider/posting_detail_provider.dart';
-import 'package:lyc_flutter_project/styles/default_padding.dart';
-import 'package:lyc_flutter_project/widget/normal_appbar.dart';
+import 'package:lyc_flutter_project/common/widget/default_padding.dart';
+import 'package:lyc_flutter_project/common/widget/normal_appbar.dart';
 import 'package:lyc_flutter_project/common/widget/round_image.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:provider/provider.dart';
