+++ lyc_flutter_project/lib/screens/find_id_screen_1.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:lyc_flutter_project/data/app_color.dart';
 import 'package:lyc_flutter_project/styles/app_text_style.dart';
-import '../widget/normal_appbar.dart';
+import '../common/widget/normal_appbar.dart';
 import 'find_id_screen_2.dart';
 
 class FindIdScreen1 extends StatelessWidget {
